Allow non-default item levels to be specified for searches
Currently, whenever an item is specified, it is always used, thus
allowing item searches to be conducted for only one or two items
instead of all three.

However, it is possible that certain items will not be the default
level, and this could be useful when determining whether or not a
certain choice is better than others.