Add Emblem optimization and sorting
Rearrange combination iteration so that the calculation can occur
inside of a function instead of creating an array to pass into a
function.  This is because the number of emblem combinations can
easily exhaust memory.

Also limit the number of sorted results to prevent exhausting
memory.

Add a stored fitness calculation since it can no longer be done
after all other processing as it is required to limit the returned
list.