Improve emblem search speeds with color requirements
By specifying a required number of colored emblems (which should
be the most common application), there is an opportunity to reduce
the complexity of the search from an n-pick-10 combination down
to c-pick-6 times (n-c)-pick-4 (or whatever the particular color
requirement is instead of six).

By simply splitting the search into two halves, the factorial size
is reduced by an order of magnitude.

Some rework had to be introduced to allow a different function to
be called when the (now given leg of the) search is finished.