Add support for emblem modifier multipliers
With the bugfixes to Lugia and Ho-oh, emblems can now have values
other than defaults for the given grades. These values are,
however, a simple 1.5 times the magnitude of the modifications at
the given grade, so introduce the ability to apply this factor
rather than record each of the stat bonuses and penalties
individually.