Rework autoattacks to be standalone moves
As the moveset list (which really should be renamed) is now Effects
instead of Moves, take the Basic and Boosted attacks out and make
them members of the Pokemon class.

Utilize the new standalone members to handle complex time
calculations instead of relying solely on the current aps value.

Add a new preprocessor field for custom auto attack Moves besides
the simple defaults.