Add structure for recording types of HP modifications
Introduce a PointStore class to record changes to health instead
of just lumping everything into a single integer.  Record whether
damage or healing or even shields are being provided and to which
target.  Use the critCalc function to multiplex this structure
into several suffixed records, with the total still being the
unsuffixed default.

It is now possible to introduce lifesteal and spellvamp while
allowing it to be ignored.