Prevent INI parsing from regressing to name-value pairs
It is possible for JavaScript code inside of learnset function
blocks to match the regex for a name=value pairing, which ends the
learnset parsing and treats the code as a custom Pokemon value.

Start the sequence parsing at an impossible (negative) index and
prevent the parsing from being able to re-enter this state. Only
accept name-value pairs before any learnset data is detected, and
once it is, allow for no state to be entered that will allow such
again.

The only cost is a check to see if the learnset sequence has
initialized when reading every line of the learnset.